restructure project to make it scalable
