+++ services/initial/functions.py
@@ -1,0 +1,40 @@
+from telegram.ext import ConversationHandler
+from telegram import ReplyKeyboardMarkup, KeyboardButton
+from services.logger import logger
+
+
+def menu(bot, update):
+    logger.info("Menu command")
+    keyboard = [
+        [KeyboardButton("Settings")]
+    ]
+    reply_markup = ReplyKeyboardMarkup(keyboard,
+                                       one_time_keyboard=False,
+                                       resize_keyboard=True)
+
+    bot.send_message(update.message.chat_id, "Select an option", reply_markup=reply_markup)
+
+
+def start(bot, update):
+    logger.info("Start command")
+    keyboard = [
+        [KeyboardButton("Menu")]
+    ]
+    reply_markup = ReplyKeyboardMarkup(keyboard,
+                                       one_time_keyboard=False,
+                                       resize_keyboard=True)
+    bot.send_message(update.message.chat_id, "Hello!", reply_markup=reply_markup)
+
+
+def settings(bot, update):
+    logger.info("Settings command")
+
+    keyboard = [
+        [KeyboardButton("Configure email receiver")],
+        [KeyboardButton("Back to menu")]
+    ]
+    reply_markup = ReplyKeyboardMarkup(keyboard,
+                                       one_time_keyboard=False,
+                                       resize_keyboard=True)
+
+    bot.send_message(update.message.chat_id, "Select an option", reply_markup=reply_markup)
